+++ src/main/java/uk/gov/hmcts/reform/civil/notification/handlers/defendantResponse/unspec/online/fulldefence/DefendantResponseRespSolTwoEmailDTOGenerator.java
@@ -1,25 +1,24 @@
-package uk.gov.hmcts.reform.civil.notification.handlers.defendantresponse;
+package uk.gov.hmcts.reform.civil.notification.handlers.defendantResponse.unspec.online.fulldefence;
 
 import org.springframework.stereotype.Component;
 import uk.gov.hmcts.reform.civil.model.CaseData;
-import uk.gov.hmcts.reform.civil.notification.handlers.RespSolOneEmailDTOGenerator;
+import uk.gov.hmcts.reform.civil.notification.handlers.RespSolTwoEmailDTOGenerator;
 import uk.gov.hmcts.reform.civil.notify.NotificationsProperties;
 import uk.gov.hmcts.reform.civil.service.OrganisationService;
 
 import java.util.Map;
 
 import static uk.gov.hmcts.reform.civil.enums.AllocatedTrack.toStringValueForEmail;
-import static uk.gov.hmcts.reform.civil.enums.MultiPartyScenario.getMultiPartyScenario;
 import static uk.gov.hmcts.reform.civil.utils.PartyUtils.getPartyNameBasedOnType;
 
 @Component
-public class DefendantResponseRespSolOneEmailDTOGenerator extends RespSolOneEmailDTOGenerator {
+public class DefendantResponseRespSolTwoEmailDTOGenerator extends RespSolTwoEmailDTOGenerator {
 
     protected static final String REFERENCE_TEMPLATE = "defendant-response-applicant-notification-%s";
 
     private final NotificationsProperties notificationsProperties;
 
-    public DefendantResponseRespSolOneEmailDTOGenerator(NotificationsProperties notificationsProperties, OrganisationService organisationService) {
+    public DefendantResponseRespSolTwoEmailDTOGenerator(NotificationsProperties notificationsProperties, OrganisationService organisationService) {
         super(organisationService);
         this.notificationsProperties = notificationsProperties;
     }
@@ -36,13 +35,9 @@ public class DefendantResponseRespSolOneEmailDTOGenerator extends RespSolOneEmai
 
     @Override
     protected Map<String, String> addCustomProperties(Map<String, String> properties, CaseData caseData) {
-        String respondentName = switch (getMultiPartyScenario(caseData)) {
-            case ONE_V_ONE, TWO_V_ONE -> getPartyNameBasedOnType(caseData.getRespondent1());
-            default -> getPartyNameBasedOnType(caseData.getRespondent1()) +
-                " and " +
-                getPartyNameBasedOnType(caseData.getRespondent2());
-        };
-        properties.put(RESPONDENT_NAME, respondentName);
+        properties.put(RESPONDENT_NAME, getPartyNameBasedOnType(caseData.getRespondent1()) +
+            " and " +
+            getPartyNameBasedOnType(caseData.getRespondent2()));
         properties.put(ALLOCATED_TRACK, toStringValueForEmail(caseData.getAllocatedTrack()));
         return properties;
     }
