+++ src/test/java/uk/gov/hmcts/reform/civil/notification/handlers/defendantResponse/unspec/offline/otherresponses/DefRespCaseOfflineAppSolOneEmailDTOGeneratorTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.civil.notification.handlers.defendantresponsecasetransferoffline;
+package uk.gov.hmcts.reform.civil.notification.handlers.defendantResponse.unspec.offline.otherresponses;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
